switch to new settlement buffer
